+++ src/Interfaces/RadioButtonInterface.php
@@ -63,6 +63,15 @@ interface RadioButtonInterface
     public function isSelected(): bool;
 
     /**
+     * Sets the label.
+     *
+     * @since 2.1.0
+     *
+     * @param string $label The label.
+     */
+    public function setLabel(string $label): void;
+
+    /**
      * Sets the name.
      *
      * @since 1.0.0